Small refactor of JsonOptions
