--- Harvest.Core/Utilities/StandardJsonOptions.cs
@@ -1,29 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Text;
-using System.Text.Json;
-using System.Text.Json.Serialization;
-
-namespace Harvest.Core.Utilities
-{
-    public static class StandardJsonOptions
-    {
-        public static JsonSerializerOptions GetOptions()
-        {
-            return new JsonSerializerOptions
-            {
-                PropertyNamingPolicy = JsonNamingPolicy.CamelCase,
-                NumberHandling = JsonNumberHandling.AllowReadingFromString,
-                AllowTrailingCommas = true,
-                PropertyNameCaseInsensitive = true,
-                ReferenceHandler = ReferenceHandler.Preserve
-            };
-        }
-
-        public static JsonSerializerOptions WithGeoJson(this JsonSerializerOptions options)
-        {
-            options.Converters.Add(new NetTopologySuite.IO.Converters.GeoJsonConverterFactory());
-            return options;
-        }
-    }
-}
