+++ Harvest.Web/Startup.cs
@@ -57,7 +57,7 @@ namespace Harvest.Web
                 options.Filters.Add<SerilogControllerActionFilter>();
             }).AddJsonOptions(options =>
             {
-                var o = StandardJsonOptions.GetOptions();
+                var o = JsonOptions.Standard;
                 options.JsonSerializerOptions.PropertyNamingPolicy = o.PropertyNamingPolicy;
                 options.JsonSerializerOptions.NumberHandling = o.NumberHandling;
                 options.JsonSerializerOptions.AllowTrailingCommas = o.AllowTrailingCommas;
@@ -153,8 +153,7 @@ namespace Harvest.Web
             services.AddScoped<INotificationService, NotificationService>();
             services.AddScoped<IEmailBodyService, EmailBodyService>();
             services.AddScoped<IEmailService, EmailService>();
-            // JsonSerializerOptions is transient so that it can be  modified as needed
-            services.AddTransient(provder => StandardJsonOptions.GetOptions());
+            services.AddScoped(provder => JsonOptions.Standard);
             services.AddScoped<IProjectHistoryService, ProjectHistoryService>();
         }
